server/requirements.txt
pipecat-ai[smallwebrtc,openai,silero,webrtc]
fastapi
uvicorn
aiohttp
numpy
python-dotenv
